structured code a bit
