+++ .gitignore
@@ -1,2 +1,4 @@
 src/binary_tree
+src/tests
 *.swp
+*.o
